server/requirements.txt
flights>=0.8.2
fastapi>=0.115.0
uvicorn>=0.32.0
